13-2. [게임 시작과 종료]플레이어 피격
<Player>
* OnCollisionEnter2D : 충돌
- Time.deltaTime -> 적절한 피격 대미지 계산
- childCount : 자식 오브젝트 개수
- GetChild : 주어진 인덱스의 자식 오브젝트 반환

+++ Assets/Undead Survivor/Codes/GameManager.cs
@@ -13,8 +13,10 @@ public class GameManager : MonoBehaviour
     public float maxGameTime = 2 * 10f; // 최대 게임 시간
     
     [Header("# Player Info")]
-    public int health;
-    public int maxHealth = 100;
+    // public int health;
+    public float health;
+    // public int maxHealth = 100;
+    public float maxHealth = 100;
     public int level; // 레벨
     public int kill; // 킬수
     public int exp; // 경험치
